#2817: updated to trunk r16140
git-svn-id: https://src.heuristiclab.com/svn/core/branches/2817-BinPackingSpeedup@16141 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ HeuristicLab.Encodings.SymbolicExpressionTreeEncoding/3.4/Compiler/Instruction.cs
@@ -28,7 +28,7 @@ namespace HeuristicLab.Encodings.SymbolicExpressionTreeEncoding {
     // op code of the function that determines what operation should be executed
     public byte opCode;
     // number of arguments of the current instruction
-    public byte nArguments;
+    public ushort nArguments;
     // an optional object value (addresses for calls, argument index for arguments)
     public object data;
   }
